build `.remove(key)` method which will remove the the key/value pair from the hashmap. add `.length()` method to return the count of nodes in the linked list. (#12)
* set up console log statements to properly test if the remove method is working. add remove method in hashmap factory as a placeholder at the moment

* expose isEmpty() method in linked list factory to be used in hashmap

the alternative would be to expose the head() node. but i feel it makes more sense just to use isEmpty, because it is more descriptive on what is going on at the hashmap method implementation level

* add further test cases for removing the other added element too, and an element that doesn't exist (these won't pass just yet as of this commit). add .length() method to find the length of the linkedList

finding the length of linked list will be helpful potentially if we check if there is only a single node in the linked list (bucket). if so, we just pop it (this method also needs to be written into the linked list factory)

* add a .removeNode method for linked list so that it removes. use the length() method in linkedlist factory to just set the head of linked list (bucket) as null if there is only a single element.

* add test case in attempt to put in the same bucket as harry but it didn't work lol. don't declare variable that won't be used in toString method for linked list. instead of pointing just to null (which disconnects the entire linked list), when we remove a node, we point the previous node to the next node's next node

* if the key is not present in the hashmap at all, we return false

i know this is kind of inefficient to then find the key again in the right bucket, but i'm not sure what to do here. i'll come back and refactor later if i have better idea

* fix test because wasn't case sensitive. instead of calling 'has' method of hashmap, i just try and find the node directly - avoids having to reinitialise and call the hash function again WITHIN the .has method of hashmap

* remove check in remove method for whether the bucket is empty

this is because this part is already within the .findNode method of the linked list as a guard clause. no need to call twice and waste time/memory