have method in linked list factory that gets the keys of all the nodes in the linked list. implement in hashmap factory too
i did this instead of using the array from getKeyValues because that already has a loop in it. instead of looping over it to get the array of arrays, and then another loop to extract the first entry of each subarray is so inefficient. i know i violate DRY a little here by having somewhat duplicate code between getKeyValues and getKeys but it's more efficient i think